requirements.txt
absl-py==1.2.0
addict==2.4.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astor==0.8.1
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==5.0.0
cachetools==4.2.4
ccimport==0.3.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cumm-cu114
cycler==0.11.0
dataclasses==0.6
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
descartes==1.1.0
docker-pycreds==0.4.0
entrypoints==0.4
fastjsonschema==2.15.3
filelock==3.6.0
filterpy==1.4.5
fire==0.4.0
fonttools==4.29.1
future==0.18.2
gast==0.2.2
gdown==4.4.0
gitdb==4.0.9
gitpython==3.1.27
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.47.0
h5py==3.7.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.7.0
ipykernel==6.13.0
ipython==7.32.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
jinja2==3.1.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.2.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.2.1
jupyterlab-widgets==1.1.0
keras-applications==1.0.8
keras-preprocessing==1.1.2
kiwisolver==1.3.2
lark==1.1.2
llvmlite==0.38.0
markdown==3.3.6
markupsafe==2.1.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
motmetrics==1.1.3
nbclient==0.6.0
nbconvert==6.5.0
nbformat==5.3.0
nest-asyncio==1.5.5
ninja==1.10.2.3
notebook==6.4.10
numba==0.55.1
numpy==1.21.6
oauthlib==3.2.0
opencv-python==4.5.5.64
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pathtools==0.1.2
pccm==0.3.4
pexpect==4.8.0
pickleshare==0.7.5
pillow==9.1.0
portalocker==2.4.0
prometheus-client==0.14.1
promise==2.3
prompt-toolkit==3.0.29
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.9.2
pycocotools==2.0.4
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.7
pyquaternion==0.9.9
pyrsistent==0.18.1
pysocks==1.7.1
python-dateutil==2.8.2
pytz==2022.1
pyyaml==5.1
pyzmq==22.3.0
qtconsole==5.3.0
qtpy==2.0.1
requests==2.28.1
requests-oauthlib==1.3.1
rsa==4.9
scikit-learn==0.19.2
scipy==1.4.1
send2trash==1.8.0
sentry-sdk==1.5.12
setproctitle==1.2.3
setuptools==62.2.0
shapely==1.8.1.post1
shortuuid==1.0.9
smmap==5.0.0
soupsieve==2.3.1
spconv-cu114
tensorboard==2.1.1
tensorflow==2.1.0
tensorflow-estimator==2.1.0
tensorflow-gpu==2.1.0
termcolor==1.1.0
terminado==0.13.3
terminaltables==3.1.10
tinycss2==1.1.1
torchinfo==1.6.6
torchsummary==1.5.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.12
wandb==0.12.17
waymo-open-dataset-tf-2-1-0==1.2.0
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.2.2
widgetsnbextension==3.6.0
wrapt==1.14.1
zipp==3.8.1